% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GB_to_hgnc.R
\name{GB_to_hgnc}
\alias{GB_to_hgnc}
\title{genBank AN to hgnc symbol}
\usage{
GB_to_hgnc(x, genBank.col, ...)
}
\arguments{
\item{x}{The table of features of an ExpressionSet or any data.frame or
matrix.}

\item{genBank.col}{The name of the column that contains the genBank
accession numbers.}
}
\value{
A \code{data.frame} containing the previous information in \code{x} plus 5
  more columns: \code{start_position} (of the gene), \code{end_position} (of
  the gene), \code{chromosome_name}, \code{hgnc_symbol} and
  \code{entrezgene}
}
\description{
Obtaining hgnc symbol and entrezgene from genBank accession number
}
\details{
Get HUGO gene nomenclature committee (hgnc) symbol from genBank accession
number, as well as the start and end position of the gene, the chromosome
and the entrezgene (NCBI gene ID)
}
