% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEDY.R
\name{getEDY}
\alias{getEDY}
\title{getEDY}
\usage{
getEDY(x, gender.var, male.key, gene.key, coef = 1.2, log = TRUE,
  group.var, control.key, experiment.type, ...)
}
\arguments{
\item{x}{An ExpressionSet or RangedSummarizedExperiment from microarray or
RNAseq experiments.}

\item{gender.var}{A string indicating the name of the column in the table of
phenotype (accessed by \code{pData} or \code{colData}) that contains the
gender of the individuals.}

\item{male.key}{A string indicating the symbol that idendtifies males (e.g.,
"male", "M", ...).}

\item{gene.key}{A string indicating the name of the column that contains the
Gene Symbol in the table of features (accessed by \code{fData} or
\code{rowData}).}

\item{coef}{Numerical. Value to consider an outlier when calling getEDY. Default
correspond to 1.2 which keeps 5\% of data in the normal case.}

\item{log}{Logical. It is set to \code{TRUE} if the gene expression values
are given in a logarithmic scale in the data set and \code{FALSE}
otherwise. Default is \code{TRUE}.}

\item{group.var}{A string indicating the name of the column that contains the
information about whether the individual is case or control (accessed by
\code{pData} or \code{colData}).}

\item{control.key}{A string indicating the symbol that identifies control
group (e.g., "control").}

\item{experiment.type}{A string indicating whether the data set is a
\strong{microarray} or a \strong{RNAseq} experiment}
}
\value{
A list containing 4 objects: 
\itemize{ 
  \item \code{EDY}: a factor with two levels, \code{NO} and \code{YES} , 
    that indicates whether the individual has EDY or not.
  \item \code{EDYcontinuous}: a vector with individual relative expression
    of chromosome Y with regard the autosomal genes.
  \item \code{threshold}: a number indicating the threshold from which down 
    an individual is considered to have EDY. 
  \item \code{eSet} or \code{RangedSummarizedExperiment}: an ExpressionSet 
    or RangedSummarizedExperiment that contains only males from the initial
    data set. It contains a new column in \code{pData} or \code{colData}
    containing the EDY status of each individual.
  }
}
\description{
Detection of individuals with Extreme Downregulation of chromosome Y (EDY)
from transcriptomic data
}
\details{
To assess EDY, the function measures the relative expression of the entire
chromosome Y with respect to the autosomes for each individual. For \emph{n}
probes (exons) in chromosome Y, with \eqn{x{i}} intensity (read counts) for
the \emph{i-th} probe, it computes \eqn{y = 1/n ·  \sumlog2(x{i}) } as a
measure of the average expression of chromosome Y (summation limits between
\emph{i=1} and \emph{N}). Likewise, for \emph{m} probes in the autosomes, the
function computes the mean expression of autosomes \eqn{a = 1/m ·
\sumlog2(x{i})} (summation limits between \emph{i=1} and \emph{M})
[\strong{NOTE:} for RNAseq data \eqn{log2(x{i} + 1)} is computed to avoid
problems with zero counts]. The relative amount of an individual's Y
expression with respect to the individual's autosomes is then \eqn{Ry = y -
a}, and, in a population sample, the individual \emph{j} is considered with
EDY if \deqn{Ry{j} < median(Ry) - 1.2 · IQR(Ry)}

where IQR is the inter-quartile range across the sample.
}
