% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEDY.R
\name{plot.EDY}
\alias{plot.EDY}
\title{Plot EDY objects}
\usage{
\method{plot}{EDY}(x, ...)
}
\arguments{
\item{x}{An EDY object generated by the function \code{getEDY}.}
}
\value{
A scatter plot in which y axis represents mean gene expression in
  chromosome Y divided by mean gene expression in the rest of the genome
  and each dot represents an individual. Individuals with extreme
  downregulation of chromosome Y are coloured in red and normal individuals
  are blue.
}
\description{
Plot EDY objects.
}
