% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printEDY.R
\name{print.EDY}
\alias{print.EDY}
\title{Print EDY objects}
\usage{
\method{print}{EDY}(x, ...)
}
\arguments{
\item{x}{An EDY object generated by the function \code{getEDY}.}
}
\value{
A description of the contents of the EDY object (\code{x}).
}
\description{
Print EDY objects.
}
