% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEDY.R
\name{getEDY}
\alias{getEDY}
\title{getEDY}
\usage{
getEDY(x, gender.var, male.key, gene.key, log = TRUE, group.var,
  control.key, ...)
}
\arguments{
\item{x}{An ExpressionSet from microarray experiment.}

\item{gender.var}{The name of the column in the table of phenotype
(\code{phenoData}) that contains the gender of the individual.}

\item{male.key}{The symbol that idendtifies males (eg.: "male", "M", ...).}

\item{gene.key}{The name of the column that contains the Gene Symbol in the
table of features (\code{featureData}).}

\item{log}{Logical. It is set to \code{TRUE} if the gene expression values
are given in a logarithmic scale in the ExpressionSet and \code{FALSE}
otherwise. Default is \code{TRUE}.}

\item{group.var}{The name of the column that contains the information about
if the individual is case or control}

\item{control.key}{The symbol that identifies control group (eg.: "control")}
}
\value{
A list containing 4 objects: 
\itemize{ 
  \item \code{Ry}: a matrix
    with the expression of each gene in chromosome Y divided by the mean
    expression of autosomal genes for each individual. 
  \item \code{EDY}: a factor with two levels, \code{YES} and \code{NO}, 
    that indicates if the individual has EDY or not.
  \item \code{threshold}: a number indicating the threshold from which down 
    an individual is considered to have EDY. 
  \item \code{eSet}: the ExpressionSet
    form the input but with the column hgnc_symbol added to featureData. 
  }
}
\description{
Extreme Downregulation of chromosome Y (EDY)
}
\details{
Detection of individuals with EDY from transcriptomic data
}
