% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEDY.R
\name{getEDY}
\alias{getEDY}
\title{getEDY}
\usage{
getEDY(x, gender.var, male.key, gene.key, coef = 1.2, log = TRUE,
  group.var, control.key, experiment.type, ...)
}
\arguments{
\item{x}{An ExpressionSet from microarray experiment.}

\item{gender.var}{A string indicating the name of the column in the table of
phenotype (accessed by \code{pData}) that contains the gender of the
individual.}

\item{male.key}{A string indicating the symbol that idendtifies males (eg.:
"male", "M", ...).}

\item{gene.key}{A string indicating the name of the column that contains the
Gene Symbol in the table of features (accessed by \code{fData}).}

\item{coef}{Numerical. Value to consider an outlier when calling LOY. Default
correspond to 1.2 which keeps 5\% of data in the normal case.}

\item{log}{Logical. It is set to \code{TRUE} if the gene expression values
are given in a logarithmic scale in the ExpressionSet and \code{FALSE}
otherwise. Default is \code{TRUE}.}

\item{group.var}{A string indicating the name of the column that contains the
information about if the individual is case or control.}

\item{control.key}{A string indicating the symbol that identifies control
group (eg.: "control").}
}
\value{
A list containing 4 objects: 
\itemize{ 
  \item \code{EDY}: a factor with two levels, \code{NO} and \code{YES} , 
    that indicates whether the individual has EDY or not.
  \item \code{EDYcontinuous}: a vector with individual relative expression
    of chromosome Y with regard the autosomal genes.
  \item \code{threshold}: a number indicating the threshold from which down 
    an individual is considered to have EDY. 
  \item \code{eSet}: an ExpressionSet that contains only males from the
  initial ExpressionSet. It contains a new column named `hgnc_symbol` in
  `fData` in case it did not have it from before, which has the HUGO gene
  nomenclature comittee symbols.
  
  }
}
\description{
Detection of individuals with Extreme Downregulation of chromosome Y (EDY)
from transcriptomic data
}
\details{
To assess EDY, we calculate the mean expression of chrY genes divided by the
mean expression of genes in the rest of the genome (for each individual).
Now, the following formulae is applied only with individuals in the control
group, or with all the individuals in case it does not exist such group:
\deqn{Threshold = median - 1.2 · IQR}
\emph{*IQR: Interquartile range}

It indicates the trheshold from which down to consider EDY.
}
