% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEDY.R
\name{plot.EDY}
\alias{plot.EDY}
\title{Plot EDY objects}
\usage{
\method{plot}{EDY}(x, ...)
}
\arguments{
\item{x}{An EDY object generated by the function \code{getEDY}.}
}
\value{
A scatter plot in which \code{y axis} represents mean gene expression
  in chromosome Y divided by mean gene expression in the rest of the genome
  and each dot represents an individual. Individuals with extreme
  downregulation of chromosome Y are coloured in light blue and labeled
  with their identifiers. Normal individuals are coloured in red. The dashed
  line indicates the threshold from which down EDY is considered.
}
\description{
Plot EDY objects.
}
