% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictEDY.R
\name{predictEDY}
\alias{predictEDY}
\title{EDY prediction}
\usage{
predictEDY(x, ...)
}
\arguments{
\item{x}{An ExpressionSet \emph{or} a matrix with CpGs in rows, samples in columns and
CpG names in the rowname \emph{or} a data.frame with individuals/samples in
columns, being the first column the CpG name.}
}
\value{
A vector with EDY status (No, Yes) or, in case 'x' is an
  ExpressionSet, the same ExpressionSet with a new column in \code{pData}
  named \code{edy.pred} which contains the EDY status prediction.
}
\description{
Prediction of Extreme Downregulation of chromosome Y (EDY) 
from methylation data
}
\details{
Predicts individuals' EDY status
}
